Build Cacheable Functions

reactFunc is a function generator that returns a function that is cacheable when combined with the use of reactArg. See example.

reactFunc(ARGV, ...)

reactArg(name)

Arguments

ARGV

A named list or vector that represent the formal arguments of the returned function.

...

Name-expression pairs that describe the reactive expressions defined in the parent environment of the returned function. The last one will be used as the returned value of the returned function. See example.

name

A symbol that corresponds to one of the names of ARGV (one of the argument name). See example.

Value

reactFunc returns a closure (function); name should be used inside the ... arguments to describe the argument as a reactive source.

Examples

## Build
rf <- reactFunc(
    ARGV = alist(x = 42, y = ),
    x = reactArg(x),
    y = reactArg(y),
    a = {
        print("Getting a()...")
        Sys.sleep(0.5)
        x() + 1
    },
    b = {
        print("Getting b()...")
        Sys.sleep(0.5)
        y() + 1
    },
    ans = {
        print("Getting ans()")
        Sys.sleep(0.5)
        a() + b()
    }
)
## Properties
#1. Definition
rf

#> function (x = 42, y) 
#> {
#>     assignedArgs <- as.list(environment())
#>     assignedArgNames <- names(assignedArgs)
#>     for (i in seq_along(assignedArgs)) {
#>         do.call(what = `[[<-`, args = list(eval(parse(text = "ReactiveSources")), 
#>             assignedArgNames[[i]], eval(parse(text = assignedArgNames[[i]]))), 
#>             envir = parent.env(environment()))
#>     }
#>     lastExprName <- tail(reactNames, 1)
#>     result <- shiny::isolate(do.call(lastExprName, args = list()))
#>     result
#> }
#> <environment: 0x6ce55a0>

#2. First run
m <- 6; n <- 9
system.time(ans <- rf(x = m, y = n))

#> [1] "Getting ans()"
#> [1] "Getting a()..."
#> [1] "Getting b()..."

#>    user  system elapsed 
#>   0.008   0.012   1.509 

ans

#> [1] 17

#3. Seconde run with the same arguments
system.time(ans <- rf(x = m, y = n))

#>    user  system elapsed 
#>   0.004   0.000   0.001 

ans

#> [1] 17

#4. Third run with an updated argument
n <- 7
system.time(ans <- rf(x = m, y = n))

#> [1] "Getting ans()"
#> [1] "Getting b()..."

#>    user  system elapsed 
#>   0.012   0.000   1.005 

ans

#> [1] 15

#5. Change the value of `x` to default
system.time(ans <- rf(y = n))

#> [1] "Getting ans()"
#> [1] "Getting a()..."

#>    user  system elapsed 
#>   0.008   0.004   1.005 

ans

#> [1] 51